- Added IC2 Compressor recipe adapter. - Provides limited compatibility for JEI transfer items. - Attempted to fix an issue where MagicPrimer#addStarlightInput and MagicPrimer#addStarlightOutput had the same method as gugu-utils causing conflicts. - Fix an issue where RequirementIngredientArray would have duplicate content. - Fix ColorableMachineTile#setMachineColor method being non-thread-safe for some components. - Fix an issue where machine controllers would sometimes crash on the client.

+++ src/main/java/kport/modularmagic/common/integration/crafttweaker/MagicPrimer.java
@@ -144,7 +144,7 @@ public class MagicPrimer {
     }
 
     @ZenMethod
-    public static RecipePrimer addStarlightInput(RecipePrimer primer, int amount) {
+    public static RecipePrimer addStarlightInput(RecipePrimer primer, float amount) {
         if (amount > 0)
             primer.appendComponent(new RequirementStarlight(IOType.INPUT, amount));
         else
@@ -154,7 +154,7 @@ public class MagicPrimer {
     }
 
     @ZenMethod
-    public static RecipePrimer addStarlightOutput(RecipePrimer primer, int amount) {
+    public static RecipePrimer addStarlightOutput(RecipePrimer primer, float amount) {
         if (amount > 0)
             primer.appendComponent(new RequirementStarlight(IOType.OUTPUT, amount));
         else
